EntityManagerで並べ替えるように修正
